Arreglado el set de tiles modificadas, arreglado el problema de los mensajes gigantes. Añadida funcionalidad a la clase Point. Refactor de HexagramTest (los nombre de las clases comienzan con mayus)
What does the warning "MessageManager queue size > 10000000" mean?
The JADE runtime has an internal queue (called OutBox) of messages to be delivered. When an agent sends a message (method send()) the message is inserted in the OutBox. A pool of threads is responsible for extracting messages from the OutBox and actually delivering them to receivers. That size of the OutBox of course depends on the number of messages waiting to be delivered and on the size of each message. When this size exceeds a given threshold (default is 10 MB) a warning is printed. This is not particularly dangerous, but of course it is a clear indication that either agents are generating really A LOT of traffic or they are sending really BIG messages. 
It is possible to increase the above threshold by means of the jade_core_messaging_MessageManager_maxqueuesize option whose value must be specified in bytes (e.g. 10000000 means 10 MB).

git-svn-id: https://forja.rediris.es/svn/cusl4-catastrof@111 8c4acde5-3c00-4ed3-8754-bc08a9928e99